WIP: moving stuf around preparing for plugins
NOTE: not buildable
