+++ apps/coreX/plugin/README.md
