+++ apps/coreX/plugin/plugin.go
