+++ base/mgr/builtin_test.go
@@ -1,4 +1,4 @@
-package pm
+package mgr
 
 import (
 	"testing"
