+++ base/mgr/handlers.go
@@ -1,6 +1,6 @@
-package pm
+package mgr
 
-import "github.com/threefoldtech/0-core/base/pm/stream"
+import "github.com/threefoldtech/0-core/base/pm"
 
 //Handler defines an interface to receiver the process manager events
 //A handler can be any object that implements one or many handle methods below
@@ -8,13 +8,13 @@ type Handler interface{}
 
 //ResultHandler receives the command result on exit
 type ResultHandler interface {
-	Result(cmd *Command, result *JobResult)
+	Result(cmd *pm.Command, result *pm.JobResult)
 }
 
 //MessageHandler gets called on the receive of each single message
 //from all commands
 type MessageHandler interface {
-	Message(*Command, *stream.Message)
+	Message(*pm.Command, *pm.Message)
 }
 
 //StatsHandler receives parsed stats messages
@@ -24,5 +24,5 @@ type StatsHandler interface {
 
 //PreHandler is called with the commands before exectution
 type PreHandler interface {
-	Pre(cmd *Command)
+	Pre(cmd *pm.Command)
 }
