+++ base/mgr/queue_test.go
@@ -1,15 +1,16 @@
-package pm
+package mgr
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
 	"time"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestQueue_Start(t *testing.T) {
 	var q Queue
 	q.Init()
-    ch := q.Channel()
+	ch := q.Channel()
 
 	if ok := assert.NotNil(t, ch); !ok {
 		t.Fatal()
@@ -19,7 +20,7 @@ func TestQueue_Start(t *testing.T) {
 func TestQueue_Push(t *testing.T) {
 	var q Queue
 	q.Init()
-    ch := q.Channel()
+	ch := q.Channel()
 	lock := make(chan int)
 	failed := true
 	go func() {
@@ -42,7 +43,7 @@ func TestQueue_Push(t *testing.T) {
 func TestQueue_PushQueued(t *testing.T) {
 	var q Queue
 	q.Init()
-    ch := q.Channel()
+	ch := q.Channel()
 
 	lock := make(chan int)
