+++ base/mgr/stream/buffer.go
