+++ base/mgr/stream/buffer_test.go
