+++ base/mgr/stream/input.go
@@ -10,6 +10,7 @@ import (
 	"sync"
 
 	logging "github.com/op/go-logging"
+	"github.com/threefoldtech/0-core/base/pm"
 )
 
 const (
@@ -30,7 +31,7 @@ var (
 
 type consumerImpl struct {
 	level   uint16
-	handler MessageHandler
+	handler pm.MessageHandler
 	source  io.Reader
 
 	multi      *bytes.Buffer
@@ -44,7 +45,7 @@ type header struct {
 }
 
 //Consume consumes a stream to the end, and calls the handler with the parsed stream messages
-func Consume(wg *sync.WaitGroup, source io.ReadCloser, level uint16, handler MessageHandler) {
+func Consume(wg *sync.WaitGroup, source io.ReadCloser, level uint16, handler pm.MessageHandler) {
 	c := &consumerImpl{
 		level:   level,
 		handler: handler,
@@ -97,8 +98,8 @@ func (c *consumerImpl) process(buffer []byte) {
 	if c.multi != nil {
 		//we are in a middle of a multi line message
 		if c.multi.Len() > 0 && c.multi.Bytes()[c.multi.Len()-1] == '\n' && bytes.HasPrefix(buffer, []byte(":::\n")) {
-			c.handler(&Message{
-				Meta:    NewMeta(c.multiLevel),
+			c.handler(&pm.Message{
+				Meta:    pm.NewMeta(c.multiLevel),
 				Message: c.multi.String(),
 			})
 			c.multi = nil
@@ -106,8 +107,8 @@ func (c *consumerImpl) process(buffer []byte) {
 		} else if end := strings.Index(string(buffer), multiLineTerm); end != -1 {
 			//we found the termination string
 			c.multi.Write(buffer[:end])
-			c.handler(&Message{
-				Meta:    NewMeta(c.multiLevel),
+			c.handler(&pm.Message{
+				Meta:    pm.NewMeta(c.multiLevel),
 				Message: c.multi.String(),
 			})
 			c.multi = nil
@@ -129,8 +130,8 @@ func (c *consumerImpl) process(buffer []byte) {
 
 		//if we reach here then we can safely flush what we have in buffer as a message
 		if i > start {
-			c.handler(&Message{
-				Meta:    NewMeta(c.level),
+			c.handler(&pm.Message{
+				Meta:    pm.NewMeta(c.level),
 				Message: string(buffer[start:i]),
 			})
 		}
@@ -147,8 +148,8 @@ func (c *consumerImpl) process(buffer []byte) {
 			} else {
 				msg = string(buffer[start : i+1]) //include the new line
 			}
-			c.handler(&Message{
-				Meta:    NewMeta(h.level),
+			c.handler(&pm.Message{
+				Meta:    pm.NewMeta(h.level),
 				Message: msg,
 			})
 
@@ -168,8 +169,8 @@ func (c *consumerImpl) process(buffer []byte) {
 		j := i + h.length
 		if end := strings.Index(string(buffer[j:]), multiLineTerm); end != -1 {
 			//we found the termination string
-			c.handler(&Message{
-				Meta:    NewMeta(h.level),
+			c.handler(&pm.Message{
+				Meta:    pm.NewMeta(h.level),
 				Message: string(buffer[j : j+end]),
 			})
 			i = j + end + len(multiLineTerm) // 4 is the width of the termination string
@@ -184,8 +185,8 @@ func (c *consumerImpl) process(buffer []byte) {
 	}
 
 	if start < len(buffer) {
-		c.handler(&Message{
-			Meta:    NewMeta(c.level),
+		c.handler(&pm.Message{
+			Meta:    pm.NewMeta(c.level),
 			Message: string(buffer[start:]),
 		})
 	}
