+++ base/mgr/stream/meta_test.go
