+++ base/mgr/testing.go
@@ -1,13 +1,17 @@
-package pm
+package mgr
 
-import "syscall"
+import (
+	"syscall"
+
+	"github.com/threefoldtech/0-core/base/pm"
+)
 
 //TestingPIDTable is used for testing to mock the process manager
 type TestingPIDTable struct{}
 
 //RegisterPID notify the process manager that a process has been started
 //with the given PID
-func (t *TestingPIDTable) RegisterPID(g GetPID) (int, error) {
+func (t *TestingPIDTable) RegisterPID(g pm.GetPID) (int, error) {
 	return g()
 }
