+++ base/pm/runnerhook.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/threefoldtech/0-core/base/pm/stream"
+	"github.com/ngaut/log"
 )
 
 //RunnerHook is a per process event handler
@@ -14,7 +14,7 @@ type RunnerHook interface {
 	//Tick is called if certain amount has passed sicne process starting
 	Tick(delay time.Duration)
 	//Message is called on each received message from the process
-	Message(msg *stream.Message)
+	Message(msg *Message)
 	//Exit is called on process exit
 	Exit(state JobState)
 	//PID is executed once the child process is started and got an PID
@@ -24,10 +24,10 @@ type RunnerHook interface {
 //NOOPHook empty handler
 type NOOPHook struct{}
 
-func (h *NOOPHook) Tick(delay time.Duration)    {}
-func (h *NOOPHook) Message(msg *stream.Message) {}
-func (h *NOOPHook) Exit(state JobState)         {}
-func (h *NOOPHook) PID(pid int)                 {}
+func (h *NOOPHook) Tick(delay time.Duration) {}
+func (h *NOOPHook) Message(msg *Message)     {}
+func (h *NOOPHook) Exit(state JobState)      {}
+func (h *NOOPHook) PID(pid int)              {}
 
 //DelayHook called after a certain amount of time passes from process start
 type DelayHook struct {
@@ -81,14 +81,14 @@ func (h *PIDHook) PID(pid int) {
 type MatchHook struct {
 	NOOPHook
 	Match  string
-	Action func(msg *stream.Message)
+	Action func(msg *Message)
 
 	io sync.Once
 	p  *regexp.Regexp
 	o  sync.Once
 }
 
-func (h *MatchHook) Message(msg *stream.Message) {
+func (h *MatchHook) Message(msg *Message) {
 	h.io.Do(func() {
 		p, e := regexp.CompilePOSIX(h.Match)
 		if e != nil {
@@ -117,14 +117,14 @@ type StreamHook struct {
 	Stderr bytes.Buffer
 }
 
-func (h *StreamHook) append(buf *bytes.Buffer, msg *stream.Message) {
+func (h *StreamHook) append(buf *bytes.Buffer, msg *Message) {
 	buf.WriteString(msg.Message)
 }
 
-func (h *StreamHook) Message(msg *stream.Message) {
-	if msg.Meta.Level() == stream.LevelStdout {
+func (h *StreamHook) Message(msg *Message) {
+	if msg.Meta.Level() == LevelStdout {
 		h.append(&h.Stdout, msg)
-	} else if msg.Meta.Level() == stream.LevelStderr {
+	} else if msg.Meta.Level() == LevelStderr {
 		h.append(&h.Stderr, msg)
 	}
